test: unit test coverage for `parse_utils.py`
